File processing
Given a file “stateinfo.txt” containing names of the state and cities
separated by “:”, create a file for each state named as “statename”.txt
containing names of cities in that state. Sample input file
“stateinfo.txt” is attached. Steps to follow:
Walk through the file. Create a dictionary whose key is the state name
and value is the file handle. Write city names into the file.
Do close all the files at the end of processing using values in
dictionary.