Pascal's triangle and threading sum
Given n, generate Pascal triangle for n rows. Use list of lists.
If n = 5, output should be
1
1 1
1 2 1
1 3 3 1
1 4 6 4 1
Write a program to sum the elements of an array (list) using 4 threads.
Let each thread add quarter of the array. Assume that the size of the
array is a multiple of 4. Use global variable “total_sum” which gets
updated by each thread once it computes partial sum. Use lock when
global variable is getting updated.